backend/user_microservice/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
mysql-connector-python==8.0.27
pydantic==1.8.2
